+++ data/sources/gitStats.ts
@@ -1,3 +1,3 @@
 export const gitStats = {
-    "commitTotal": "7943"
+    "commitTotal": "7949"
 }