+++ packages/remark-typography/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.6](https://github.com/codsen/codsen/compare/remark-typography@0.4.5...remark-typography@0.4.6) (2022-10-31)
+
+**Note:** Version bump only for package remark-typography
+
 ## [0.4.4](https://github.com/codsen/codsen/compare/remark-typography@0.4.3...remark-typography@0.4.4) (2022-10-22)
 
 ### Fixed